requirements.txt
pyserial
psutil

pypiwin32
wmi
pyadl